+++ pytorch_lightning/trainer/connectors/logger_connector/logger_connector.py
@@ -269,6 +269,11 @@ class LoggerConnector:
         return is_different_fx and is_first_batch
 
     def reset(self, metrics: Optional[bool] = None) -> None:
+        if self.trainer.sanity_checking:
+            # reset metrics
+            self._progress_bar_metrics = {}
+            self._logged_metrics = {}
+            self._callback_metrics = {}
         self.trainer._results.reset(metrics=metrics)
         self._batch_idx = None
         self._split_idx = None